Start extracting common code into a separate repo
