+++ README.md
@@ -6,6 +6,7 @@ OpenDota Web UI: A web interface for viewing Dota 2 data. This utilizes the [Ope
 
 Quickstart
 ----
+* Clone the repo using `git clone --recurse-submodules git://github.com/odota/web`
 * Install Node.js (6.0.0 or greater) (on Ubuntu, `curl -sL https://deb.nodesource.com/setup_8.x | sudo -E bash - && sudo apt-get install -y nodejs`)
 * `npm install`
 * `npm start`
