+++ public/index.html
@@ -123,7 +123,6 @@
 </head>
 
 <body>
-
   <div id="root" />
   <noscript>
     <h3>Looks like your browser isn't allowing scripts! Enable JavaScript to run this app.</h3>
@@ -139,25 +138,5 @@
     <div class="sk-cube sk-cube8"></div>
     <div class="sk-cube sk-cube9"></div>
   </div>
-  </div>
-  <script>
-    if ('serviceWorker' in navigator) {
-        /*
-        navigator.serviceWorker
-          .register('/sw.js')
-          .then(function() {
-            console.log('Service worker registered!');
-          })
-          .catch(function(err) {
-            console.log(err);
-          });
-          navigator.serviceWorker.getRegistrations().then(function(registrations) {
-            registrations.forEach(function (r) {
-              r.unregister();
-            });
-          });
-        */
-      }  
-  </script>   
 </body>
 </html>
