--- src/components/App/AppLogo.spec.jsx
@@ -1,16 +1,0 @@
-/*
-import React from 'react';
-import { expect } from 'chai';
-import { shallow } from 'enzyme';
-import { Link } from 'react-router-dom';
-
-import AppLogo from './AppLogo';
-
-describe('AppLogo tests', () => {
-  it('should render the app name inside of a <Link>', () => {
-    const ShallowAppLogo = shallow(<AppLogo />);
-    const AppLogoLink = ShallowAppLogo.find(Link);
-    expect(AppLogoLink).to.exist();
-  });
-});
-*/
